pyproject.toml
[project]
name = "uht-substrate"
version = "1.0.0"
description = "UHT Substrate Agent - MCP server using Universal Hex Taxonomy as ground truth"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "UHT Substrate Team" }]

dependencies = [
    # MCP Server Framework
    "fastmcp>=2.0.0",

    # Async HTTP client for UHT API
    "httpx>=0.27.0",

    # Data validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Neo4j graph database
    "neo4j>=5.15.0",

    # YAML for priors configuration
    "pyyaml>=6.0.1",

    # Caching
    "cachetools>=5.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "uuid6>=2024.1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
uht-substrate = "uht_substrate.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uht_substrate"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
